stopping scan only if bt adapter is still ON
If user disables bluetooth while scanning is in progress, then app will hard crash when attempting to stop the scan, so a check was introduced for it. The "BleManagerStopScan" event is still sent, however, because there are bound to be important actions in listeners for the stopping of the scan, so at the end of the scanning period these actions can still be performed (even though the scan was not under way since the user disabled bluetooth).

This is not a perfect solution, but is an improvement over hard crash. A better solution would be: when BluetoothAdapter changes state to STATE_OFF, check whether scan is ongoing, and if it is, then stop it. I will open an issue on this for visibility since I don't have time to implement it the "perfect way" now, and anyone can take it from there or I might once I get the chance.