Implement companion device manager support
Android's `CompanionDeviceManager` is used to get access[1,2] (association)
to a Bluetooth LE (and classic Bluetooth) devices without requiring too
broad permission set (notably location on older android versions).

Using `CompanionDeviceManager` is also a requirement for implementing
background connection management, tho' it also requires the application
to implement `CompationDeviceService`[3].

`CompationDeviceManager` availability is not guaranteed, hence the few
API version and feature detection checks.

Companion scanning is intentionally behind different API entry (`scan`
vs `companionScan`). This is because on Android you might end up needing
to scan for a peripheral that you already know as a companion. This
happens when you have a companion that is not bonded.

iOS stubs are missing at the moment.

1: https://developer.android.com/develop/connectivity/bluetooth/companion-device-pairing
2: https://developer.android.com/reference/android/companion/CompanionDeviceManager
3: https://developer.android.com/reference/android/companion/CompanionDeviceService